Improved gift subscriptions end-to-end tests
no issue
